+++ src/pyobs/improve.py
@@ -4,7 +4,6 @@
 
 import os
 import sys
-from types    import *
 from numpy    import loadtxt, ones, savez, pi, log, concatenate, arange
 from datetime import timedelta,datetime as TIME
 
@@ -44,7 +43,7 @@ class IMPROVE(object):
 
         # Read each granule, appending them to the list
         # ---------------------------------------------
-        if type(Path) is ListType:
+        if type(Path) is list:
             if len(Path) == 0:
                 print("WARNING: Empty MxD04_L2 object created")
                 return
@@ -218,7 +217,7 @@ class SITE_MAP(object):
 
         # Read each granule, appending them to the list
         # ---------------------------------------------
-        if type(Path) is ListType:
+        if type(Path) is list:
             if len(Path) == 0:
                 print("WARNING: Empty MxD04_L2 object created")
                 return